% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fermat.test}
\alias{fermat.test}
\title{Test an integer for primality with Fermat's little theorem.}
\usage{
fermat.test(n)
}
\arguments{
\item{n}{the integer to test for primality}
}
\value{
Whether the integer passes the Fermat test
}
\description{
Test an integer for primality with Fermat's little theorem.
}

