% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.US.data.R
\name{get.US.data}
\alias{get.US.data}
\title{Get US data for estimation procedure}
\usage{
get.US.data(folder = "data/input/", file = "rstar.data.us.csv")
}
\arguments{
\item{folder}{the csv to perform the HLW estimation on. Please see elp file for required column names}

\item{file}{a string for the country you are performing the estimates on - this is used in the output file name}
}
\description{
Downloads US data from FRED website and saves as a dataframe
funtion takes 2 arguments - the output folder and filename
}
\examples{
folder <-  "data/inputData/"         # output folder
file <-  "rstar.data.us.csv"     # output filename

get.US.data(folder = folder, file = file)
}
