% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.hlw.R
\name{hlw_nri}
\alias{hlw_nri}
\title{Run estimation on specified file}
\usage{
hlw_nri(input_file, country_name, output_folder, niter = 5000,
  run.se = TRUE)
}
\arguments{
\item{input_file}{the csv to perform the HLW estimation on. Please see elp file for required column names}

\item{country_name}{a string for the country you are performing the estimates on - this is used in the output file name}

\item{output_folder}{String}

\item{niter}{Integer. Sets the number of iterations for Monte Carlo Simulation. default is set to 5000}

\item{run.se}{logical. default is TRUE, set to FALSE if you do not want to compute standard errors for the estimation}
}
\description{
Takes csv with columns date, gdp.log, inflation, inflation.expectations, interest.rate as input.
The sample must be at least 98 quarters.
Saves two output files to specified folder and you should specify country name
}
\examples{
input_file <-  "data/input/rstar.data.us.test.csv"  # location of input file
country_name <-  "US.test"                     #country name for output file
output_folder <-  "data/output/"                  # output folder for estimation
run.se <-  FALSE                                    # turn on standard errors
niter <-5000       #iterations for monte carlo simulation

hlw_nri(input_file = input_file,
       country_name = country_name,
       output_folder = output_folder,
       run.se = run.se,
       niter = niter)
}
