% Generated by roxygen2 (4.0.1): do not edit by hand
\name{log_logistic}
\alias{log_logistic}
\title{Return values of log logistic function given vector of sizes}
\usage{
log_logistic(size, mu, sd)
}
\arguments{
\item{size}{a vector of (n) size values}

\item{mu}{the mean of log logistic function}

\item{sd}{the standard deviation of the log logistic function}
}
\value{
a vector of (n) values of the log logistic function corresponding to each of (n) size values
}
\description{
Return values of log logistic function given vector of sizes
}

