% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{rstar}
\alias{rstar}
\alias{rstar-package}
\title{Rstar: R Stock Assessment Routines}
\description{
The Rstar package provides complimentary code to the Cstar Library
for ADMB. Each of the common stock assessment functions available in the
Cstar Library is also available as an R function for testing purposes.
Users of Cstar functions might find it helpful to first experiment with
parameter values for any particular function by using this R package.
}
\details{
Extra details will go here.
}

